Reorganize folders for better clarity in frontend connection and external APIs
In this commit, the folder structure has been reorganized to improve clarity regarding the connection with the frontend and other external APIs. This change aims to make it easier to locate and understand the code related to these functionalities, enhancing the maintainability and development of the system.

- Created specific folders for frontend connection and external APIs
- Moved files related to these areas to their respective folders
- Updated import paths in files according to the new folder structure
- Conducted tests to ensure the correct organization and functionality of the code after the reorganization

Resolves: #11
