+++ src/app.module.ts
@@ -1,10 +1,9 @@
+import { LoginModule } from './modules/login.module';
+import { PrismaService } from 'src/infra/prisma/prisma.service';
 import { Module } from '@nestjs/common';
 import { CacheModule } from '@nestjs/cache-manager';
 // import { AppController } from './app.controller';
 // import { AppService } from './app.service';
-import { PrismaService } from './prisma/prisma.service';
-import { SignupModule } from './signup/signup.module';
-import { LoginModule } from './login/login.module';
 import * as redisStore from 'cache-manager-redis-store';
 
 @Module({
@@ -15,7 +14,6 @@ import * as redisStore from 'cache-manager-redis-store';
       host: 'redis',
       port: 6379,
     }),
-    SignupModule,
     LoginModule,
   ],
   providers: [PrismaService],
