--- src/login/login.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { LoginService } from './login.service';
-import { LoginController } from './login.controller';
-import { PrismaService } from 'src/prisma/prisma.service';
-
-@Module({
-  controllers: [LoginController],
-  providers: [LoginService, PrismaService],
-})
-export class LoginModule {}
