+++ src/modules/login.module.ts
@@ -1,0 +1,10 @@
+import { PrismaService } from '../infra/prisma/prisma.service';
+import { LoginController } from './../auth/controllers/auth.controller';
+import { LoginService } from './../services/login.service';
+import { Module } from '@nestjs/common';
+
+@Module({
+  controllers: [LoginController],
+  providers: [LoginService, PrismaService],
+})
+export class LoginModule {}
