--- src/signup/signup.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { SignupController } from './signup.controller';
-import { SignupService } from './signup.service';
-
-describe('SignupController', () => {
-  let controller: SignupController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [SignupController],
-      providers: [SignupService],
-    }).compile();
-
-    controller = module.get<SignupController>(SignupController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
