--- src/signup/signup.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { SignupService } from './signup.service';
-
-@Controller('signup')
-export class SignupController {
-  constructor(private readonly signupService: SignupService) {}
-}
