--- src/signup/signup.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { SignupService } from './signup.service';
-import { SignupController } from './signup.controller';
-
-@Module({
-  controllers: [SignupController],
-  providers: [SignupService],
-})
-export class SignupModule {}
