--- src/signup/signup.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { SignupService } from './signup.service';
-
-describe('SignupService', () => {
-  let service: SignupService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [SignupService],
-    }).compile();
-
-    service = module.get<SignupService>(SignupService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
