--- src/signup/signup.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common';
-// import { PrismaService } from 'src/prisma/prisma.service';
-
-@Injectable()
-export class SignupService {
-  //   constructor(prisma: PrismaService) {}
-  //async signup(data)
-}
